Refactor main component
